switch with label update, charts page update
switch can now have text from both sides
charts slider implemented, can choose how many samples to show, can set middle point of chart, time chart available, freq chart available(fft from time samples), time samples and freq samples can be exported to excell, formatting shown labels and values in charts,

+++ BrainWaves/BrainWaves/BrainWaves/Views/ChartsPage.xaml.cs
@@ -14,7 +14,7 @@ namespace BrainWaves.Views
             BindingContext = new ChartsViewModel();
         }
 
-        public ChartsPage(List<float> _samples)
+        public ChartsPage(List<double> _samples)
         {
             InitializeComponent();
             BindingContext = new ChartsViewModel(_samples);
